Perpendicular line test: No iterative circles
Instead of translating Koby's simulation from Python which has its own issues with compilation time and finding the optimal circle iteratively, I've implemented a different largest circle criteria.

After finding the parameters of the arm and its polytope of feasible force sets, the program then calculates the shortest distances (perpendicular line in 2D) from the center of the circle to the edges of the polytope. The smallest perpendicular line distance, that's within the polytope, is the radius of the largest circle possible in the polytope.